requirements.txt
Django>=4.2
djangorestframework
python-dotenv
redis
djangorestframework-simplejwt
